% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jammglm.f.R
\name{jammGLM}
\alias{jammGLM}
\title{GLM Mediation Model}
\usage{
jammGLM(
  data,
  dep = NULL,
  mediators = NULL,
  factors = NULL,
  covs = NULL,
  modelTerms = NULL,
  ciType = "standard",
  ciWidth = 95,
  bootN = 1000,
  contrasts = NULL,
  showRealNames = TRUE,
  showContrastCode = FALSE,
  bogus = FALSE,
  simpleScale = "mean_sd",
  cvalue = 1,
  percvalue = 25,
  simpleScaleLabels = "labels",
  scaling = NULL,
  tableOptions = list("beta", "component"),
  pathOptions = list("suggested"),
  mediatorsTerms = list(list()),
  moderatorsTerms = list(list()),
  formula
)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{a string naming the dependent variable from \code{data},
variable must be numeric. Not useful if formula is used.}

\item{mediators}{a vector of strings naming the mediators from \code{data}. Not useful if formula is used.}

\item{factors}{a vector of strings naming the fixed factors from
\code{data}. Not useful if formula is used or the variable is already a factor in the data}

\item{covs}{a vector of strings naming the covariates from \code{data}. Not useful if formula is used.}

\item{modelTerms}{a list of character vectors describing fixed effects  terms. Not useful if formula is used}

\item{ciType}{Choose the confidence interval type}

\item{ciWidth}{a number between 50 and 99.9 (default: 95) specifying the
confidence interval width for the parameter estimates}

\item{bootN}{number of bootstrap samples for estimating confidence
intervals}

\item{contrasts}{a named vector of the form \code{c(var1='type', var2='type2')} specifying the type of contrast to use,
one of \code{'deviation'}, \code{'simple'}, \code{'dummy'}, \code{'difference'}, \code{'helmert'}, \code{'repeated'} or \code{'polynomial'}.
 If NULL, \code{'simple'} is used. Can also be passed as a list of list of the form \code{list(list(var='var1',type='type1'))}.}

\item{showRealNames}{\code{TRUE} or \code{FALSE} (default), provide raw
names of the contrasts variables}

\item{showContrastCode}{\code{TRUE} or \code{FALSE} (default), provide
contrast coefficients tables}

\item{bogus}{\code{a bogus option to define a label without visible
children, used for internal checks}}

\item{simpleScale}{\code{'mean_sd'} (default), \code{'custom'} , or
\code{'custom_percent'}. Use to condition the covariates (if any).}

\item{cvalue}{offset value for conditioning. Values are mean +/- cvalue.}

\item{percvalue}{offset value for conditioning. Values are median +/- pecvalue}

\item{simpleScaleLabels}{style for presenting condition values of a moderator. It can be \code{'labels'} (default), \code{'values'} or \code{'labels_values'} for both.}

\item{scaling}{a named vector of the form \code{c(var1='type1',var2='type2')}. Types are
\code{'centered'} to the mean, \code{'standardized'}, log-transformed \code{'log'} or \code{'none'}.
\code{'none'} leaves the variable as it is. It can also be passed as a list of lists.}

\item{tableOptions}{a vector of options to be shown in the tables. \code{'component'} shows the indirect effects 
components in the results tables. \code{'regression'} show the results of all the regression (R-squared, F-test and coefficients) involved in the model.}

\item{pathOptions}{.}

\item{mediatorsTerms}{a list of lists specifying the models for with the
mediators as dependent variables. Not required in formula is used.}

\item{moderatorsTerms}{a named vector of the form c("med"=c("mod1",mod2")) specifying the moderator(s) of each mediator. 
This is required to decide for which variable we need to condition the mediated effects and to single out moderators
in the path diagram. If not specified, any interaction is considered as any other term in the model}

\item{formula}{a list of formulas to use, see the examples}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$info} \tab \tab \tab \tab \tab a table \cr
  \code{results$pathmodelgroup$pathmodel} \tab \tab \tab \tab \tab a path model \cr
  \code{results$pathmodelgroup$pathnotes} \tab \tab \tab \tab \tab a table \cr
  \code{results$models$moderationEffects} \tab \tab \tab \tab \tab a table \cr
  \code{results$models$main} \tab \tab \tab \tab \tab a table \cr
  \code{results$models$contrastCodeTables} \tab \tab \tab \tab \tab an array of contrast coefficients tables \cr
  \code{results$regressions$overall} \tab \tab \tab \tab \tab a group \cr
  \code{results$regressions$mediator_regressions} \tab \tab \tab \tab \tab an array of regressions for the mediators \cr
  \code{results$regressions$full} \tab \tab \tab \tab \tab a group \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$info$asDF}

\code{as.data.frame(results$info)}
}
\description{
GLM mediation model
}
