% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jammglm.h.R
\name{jammGLM}
\alias{jammGLM}
\title{GLM Mediation Model}
\usage{
jammGLM(data, dep = NULL, mediators = NULL, factors = NULL,
  covs = NULL, modelTerms = NULL, ciType = "standard",
  ciWidth = 95, bootN = 1000, contrasts = NULL,
  showRealNames = TRUE, showContrastCode = FALSE, bogus = FALSE,
  simpleScale = "mean_sd", cvalue = 1, percvalue = 25,
  simpleScaleLabels = "labels", scaling = NULL,
  tableOptions = list("beta"), mediatorsTerms = list(list()),
  moderatorsTerms = list(list()))
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{a string naming the dependent variable from \code{data},
variable must be numeric}

\item{mediators}{a vector of strings naming the mediators from \code{data}}

\item{factors}{a vector of strings naming the fixed factors from
\code{data}}

\item{covs}{a vector of strings naming the covariates from \code{data}}

\item{modelTerms}{a list of character vectors describing fixed effects
terms}

\item{ciType}{Choose the confidence interval type}

\item{ciWidth}{a number between 50 and 99.9 (default: 95) specifying the
confidence interval width for the parameter estimates}

\item{bootN}{number of bootstrap samples for estimating confidence
intervals}

\item{contrasts}{a list of lists specifying the factor and type of contrast
to use, one of \code{'deviation'}, \code{'simple'}, \code{'difference'},
\code{'helmert'}, \code{'repeated'} or \code{'polynomial'}}

\item{showRealNames}{\code{TRUE} or \code{FALSE} (default), provide raw
names of the contrasts variables}

\item{showContrastCode}{\code{TRUE} or \code{FALSE} (default), provide
contrast coefficients tables}

\item{bogus}{\code{a bogus option to define a label without visible
children}}

\item{simpleScale}{\code{'mean_sd'} (default), \code{'custom'} , or
\code{'custom_percent'}. Use to condition the covariates (if any)}

\item{cvalue}{offset value for conditioning}

\item{percvalue}{offset value for conditioning}

\item{simpleScaleLabels}{.}

\item{scaling}{a list of lists specifying the covariates scaling, one of
\code{'centered to the mean'}, \code{'standardized'}, or \code{'none'}.
\code{'none'} leaves the variable as it is}

\item{tableOptions}{.}

\item{mediatorsTerms}{a list of lists specifying the models for with the
mediators as dependent variables.}

\item{moderatorsTerms}{a list of lists specifying the the IV which
moderatorate each mediated effect.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$info} \tab \tab \tab \tab \tab a table \cr
  \code{results$pathmodelgroup$pathmodel} \tab \tab \tab \tab \tab a path model \cr
  \code{results$pathmodelgroup$pathnotes} \tab \tab \tab \tab \tab a table \cr
  \code{results$models$moderationEffects} \tab \tab \tab \tab \tab a table \cr
  \code{results$models$main} \tab \tab \tab \tab \tab a table \cr
  \code{results$models$contrastCodeTables} \tab \tab \tab \tab \tab an array of contrast coefficients tables \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$info$asDF}

\code{as.data.frame(results$info)}
}
\description{
GLM mediation model
}
